feat: MemberController Auth, Relation, Coordinate로 변경
